refactor: finalize CLI migration and task logging
