+++ usage/workflow_download.py
@@ -1,5 +1,5 @@
 from jmcomic import *
-from jmcomic.cl import JmcomicUI
+from jmcomic.cli import JmcomicUI
 
 # 下方填入你要下载的本子的id，一行一个，每行的首尾可以有空白字符
 jm_albums = '''
